Fix all compiler warnings related to hashcons and canonical types (#48)
* Fixed clang compiler warnings/notes

* Fixed all compiler warnings

* Better error messages

* Trimmed spaces

+++ analyze/canonical-type.h
@@ -85,7 +85,7 @@ Declaration canonical_type_decl(CanonicalType *canonical_type);
  * @arg untyped CanonicalType
  * @return hash value
  */
-int canonical_type_hash(void *arg);
+long canonical_type_hash(void *arg);
 
 /**
  * Given a canonical type, prints it to the file output
